Follow redirect when testing URL, and fix url testing.
From my understanding of the intersphinx mapping the testing was
incorrect.

Here I am a bit stricter by forcing the second parameter to be a full
url (or None), in which case the second parameter must be a subpatch of
the first one.

We also follow redirect (301) to make sure the final version exists,
otherwise 301 does not raise for status.